+++ src/Components/Blogs/BlogSingle.js
@@ -15,16 +15,28 @@ const BlogSingle = () => {
 
                 </div>
                 <article className="flex my-3 flex-col md:flex-row md:items-start space-x-6 ml-10 md:ml-0">
-                    <img loading='lazy' src='https://media.geeksforgeeks.org/wp-content/cdn-uploads/20220509120600/Learn-Data-Structures-and-Algorithms-Easily.gif' alt="" className="flex-none w-60 border-2  rounded-md bg-slate-100" />
+                    <img loading='lazy' src='https://miro.medium.com/max/875/1*fCtiD4qEZXcB5jiYPmH8YQ.png' alt="" className="flex-none w-60 border-2  rounded-md bg-slate-100" />
                     <div className="min-w-0 relative flex-auto">
 
-                        <h2 className="font-semibold text-slate-900 truncate">{'title'}</h2>
-                        <p className='mr-28'>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Vel laboriosam fugiat cupiditate nihil ex optio ab adipisci atque, earum dolorum voluptatem assumenda dignissimos, explicabo voluptas est quasi voluptates doloremque iste!</p>
+                        <h2 className="font-semibold text-slate-900 truncate"> Peer dependencies in a Node module</h2>
+                        <p className='mr-28'>In some package.json files, you might see a few lines like this:
+
+{
+    `{
+        //...
+        "peerDependencies": {
+          "libraryName": "1.x"
+        }
+      }`
+}
+You might have already seen dependencies and devDependencies, but not peerDependencies.
+
+dependencies are the packages your project depends on.s</p>
 
                         <dl className="mt-2 flex flex-wrap text-sm leading-6 font-medium">
 
                             <div className="flex-none w-full mt-2 font-normal mt-5">
-                                <a href={`/$`}> <button className='btn btn-xs border-none bg-[brown]'>Detail</button>
+                                <a href={`/blogs/blog-detail`}> <button className='btn btn-xs border-none bg-[brown]'>Detail</button>
                                 </a>
                             </div>
                         </dl>
